refactor: change naming from lecturer to academic teacher
